Very first pass at detecting straight paths through loops
This isn't really adequate or useful, but it's a first pass to get us
started.

The goal is to detect subgraphs of the DFA where we're matching a set of
constant-length strings, and determine if matching these subgraphs could
be done with a faster/simpler algorithm like Boyer-Moore.

We need a few things:
1) gather the symbols during transitions through the loop to build up
any strings that the loop is matching

2) how many exits are there from the loop?  can we classify them?
   - exit to parent loop
   - exit to sibling loop
   - exit to something unrelated?

3) many non-straight paths through the loop can be fine.
   consider /abc|def/ (unanchored).  This has an outer loop where the
   regexp forks into two "straight paths":
      S1 -[a]-> S2 -[b]-> S3 -[c] -> S4
   and:
      S1 -[d]-> S5 -[e]-> S6 -[f] -> S4

   each of the states S2, S3, S5, and S6 have back edges
   to S1, S2, and S5.

   In the simple classification, loops S2/S3 and S5/S6 are
   "straight path" while the outer loop with head S1 is not
   because of the fork.  We need a way to identify that S1
   is basically matching either "abc" or "def", both of which
   are fixed-length strings.

   Ideally, we'd also match more complex topographies that only
   represent fixed-length strings.

A thought on 3):

- the edges present are all back edges to detect non-matching
prefixes.  eg: aaaaabc, or adeabddef.  Both end in a match, but have
non-matching prefixes.
- this is different from internal loops that can extend the length of
the string.
- I'm not entirely sure how to separate these two classifications,
though.  Contrast with something like: /ab+c|def/.  Here there will be
an internal loop with the 'b+' part that isn't prefix-matching.
However, it's not clear how to distinguish this from the prefix-matching
cases.  This is the crux of the problem, I think.